refactor: Remove call to Python container
This is with the understanding that docker-compose is an orchestrator platform in the vein of kubernetes and that the stuff trying to be done with respect to calling the module will be better done in a Dockerfile.

Currently Python module cannot connect to neo4j docker container because of hosting. Need to specify the Docker container's IP address in the module.